Grouping icon search dependencies
